feat(nix-gc): clean up 0-size drv files in under /nix/store
if the machine runs low on disk space it's possible for derivation files
to be created but never get content which results in derivations that
can't be removed by the gc. this is a workaround which finds and deletes
those problem derivations.

the current implementation isn't clean because of multiple aspects which
can be improved subsequently. namely:

* no usage of platform conditional. instead using implicit knowledge
  about the configs.
* the shared cleanup command is duplicated
* on darwin, the nix-gc command is duplicated from the nix-darwin repo