feat: add dev-garage module for future reference
intermediately i used garage as an S3 storage.
it didn't fit the workflow well because garage doesn't currently support
anonymous access to S3 objects.
`rfs pack` stores the s3 credentials that it uses for pushing to the
store into the resulting flist (by default with a stripped password).

options for making this work are
* adding anonymous download support to garage
* creating a read-only credential and either modify rfs to store
  alternative credentials or post-process the flist (sqlite3 db).